[TASK] Move all update scripts into Backend folder
