+++ Classes/Controller/Backend/Update/AbstractUpdateController.php
@@ -1,5 +1,5 @@
 <?php
-namespace Ppi\TemplaVoilaPlus\Controller\Update;
+namespace Ppi\TemplaVoilaPlus\Controller\Backend\Update;
 
 /*
  * This file is part of the TYPO3 CMS project.
